# routes/swagger.yaml
openapi: 3.0.0
info:
  title: PocketMoney API
  version: 1.0.0
  description: API for PocketMoney application
servers:
  - url: 'http://localhost:3001'
    description: 'localhost'
  - url: 'http://34.170.36.196/api/'
    description: 'Google cloud'
paths:
...
  /budget/update/{id}:
    put:
      summary: Update existing budget
      description:
      tags:
        - budget
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the budget to be updated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  description: Name of the budget to be updated
      responses:
        200:
          description: Success
  /budget/delete/{id}:
    delete:
      summary: Delete budget
      description:
      tags:
        - budget
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the budget to be deleted
      responses:
        200:
          description: Success
  /budget/balance/add/{budgetid}:
    put:
      summary: Add new balance to the budget
      tags:
        - budget/balance
      parameters:
        - in: path
          name: budgetid
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the budget into which is the balance being added
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                account:
                  type: string
                  description: Balance account
                  enum: [Unicredit, Revolut, Kreditka, Hotovost]
                currentBalance:
                  type: integer
                  description: Actual balance on the account
                plannedBalance:
                  type: integer
                  description: Account balance which was planned at the beginning of the budget period
      responses:
        200:
          description: Success
  /budget/balance/update/{budgetid}:
    put:
      summary: Update budget balance
      tags:
        - budget/balance
      parameters:
        - in: path
          name: budgetid
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the budget in which the balance is being updated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                balanceId:
                  type: string
                  description: ID of the balance being updated
                balanceName:
                  type: string
                  description: Account of the balance
                  enum: [Unicredit, Revolut, Kreditka, Hotovost]
                currentBalance:
                  type: integer
                  description: Actual account balance
                plannedBalance:
                  type: integer
                  description: Actual balance planned at the beginning of the budget period
      responses:
        200:
          description: Success
  /budget/balance/delete/{budgetid}:
    put:
      summary: Delete budget balance
      tags:
        - budget/balance
      parameters:
        - in: path
          name: budgetid
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the budget in which the balance is being deleted
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                balanceId:
                  type: string
                  description: ID of the balance being deleted
      responses:
        200:
          description: Success
...
  /accounts/{id}:
    get:
      summary: Returns account detail
      tags:
        - accounts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: Account ID
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
  /accounts/new:
    post:
      summary: Add new account
      tags:
        - accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  description: Name of the account
                type:
                  type: string
                  description: Type of the account (enum)
                currency:
                  type: string
                  description: Enum
                balance:
                  type: string
                  description:
      responses:
        200:
          description: Success
  /accounts/update/{id}:
    put:
      summary: Updates account
      tags:
        - accounts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#components/schemas/Account'
      responses:
        200:
          description: Success
        400:
          description: Wrong ID in path
  /accounts/delete/{id}:
    delete:
      summary: Deletes account
      tags:
        - accounts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the account being deleted
      responses:
        200:
          description: Success
        400:
          description: Wrong ID in path
        500:
          description: Error when deleting account
components:
  schemas:
...
    Account:
      type: object
      description:
        - It is any asset that holds balance and can be used for spendings (e.g. current account, cash, credit card ...)
        - Accounts are used in budget as the holders of balances
      properties:
        _id:
          type: string
        name:
          type: string
        type:
          type: string
          description: Account type (e.g. current, savings ...)
        balance:
          type: integer
        currency:
          type: string
...